chore: refactor helper functions
