+++ src/utils/email/microsoft_flag_requests.py
@@ -1,8 +1,9 @@
 import json
 
 from ..param_types import *
-from ..helpers import *
+from ..helper_functions.helpers_email import *
 from ..token_manager import TokenManager
+from ..helper_functions.general_helpers import handle_microsoft_errors, microsoft_patch
 
 
 class MicrosoftFlagRequests:
